requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
click>=8.1.0
pandas>=2.0.0
numpy>=1.26.0
tabulate>=0.9.0
rich>=13.7.0
asciichartpy>=1.5.25
mplfinance>=0.12.9b0
matplotlib>=3.8.0
requests>=2.31.0
